Improved Exports for @sei-js/evm
Due to inconsistencies in node versions and the way they resolve this package, this change removes the "export" field in the package.json so that default module resolution strategies are used by using the "main", "module", and "types" fields in the package.json.

all files from the "viem" and "ethers" directories are included in the default export and the dependencies are moved from peer dependencies to the regular dependencies field.